Kernel: Xiaomi kernel changes for MI 8, MIX 2S, MI POCO F1, MI 8UD, MIX 3, MI 8 Explorer Edition Android Q.
The Patch based on QualComm release TAG: LA.UM.8.3.r1-05800-sdm845.0
The kernel config file used are dipper_user_defconfig,polaris_user_defconfig,beryllium_user_defconfig,
equuleus_user_defconfig,perseus_user_defconfig and ursa_user_defconfig.

+++ drivers/media/platform/msm/camera/cam_jpeg/jpeg_hw/cam_jpeg_hw_mgr.h
@@ -1,4 +1,5 @@
 /* Copyright (c) 2017-2018, The Linux Foundation. All rights reserved.
+ * Copyright (C) 2019 XiaoMi, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -38,7 +39,7 @@
 struct cam_jpeg_process_frame_work_data_t {
 	uint32_t type;
 	void *data;
-	uintptr_t request_id;
+	uint64_t request_id;
 };
 
 /**
@@ -81,7 +82,7 @@ struct cam_jpeg_hw_cfg_req {
 	struct list_head list;
 	struct cam_hw_config_args hw_cfg_args;
 	uint32_t dev_type;
-	uintptr_t req_id;
+	int64_t req_id;
 };
 
 /**
